+++ src/main/java/com/sys/park/app/dtos/Ticket/TicketDto.java
@@ -6,9 +6,9 @@ import lombok.Data;
 
 @Data
 public class TicketDto {
-    private Integer id;
-    private Integer idCustomerVehicle;
-    private Integer idVacancy;
+    private Long id;
+    private Long idCustomerVehicle;
+    private Long idVacancy;
     private LocalDateTime entryTime;
     private LocalDateTime exitTime;
     private Double totalPrice;
