+++ src/main/java/com/sys/park/app/dtos/Ticket/TicketMensalForm.java
@@ -5,17 +5,17 @@ import lombok.Data;
 
 @Data
 public class TicketMensalForm {
-    private Integer id;
+    private Long id;
 
     @NotNull(message = "A vaga não pode ser nula!")
-    private Integer idVacancy;
+    private Long idVacancy;
 
     @NotNull(message = "O tipo de cliente não pode ser nulo!")
-    private Integer idCustomerType;
+    private Long idCustomerType;
 
     @NotNull(message = "O veículo não pode ser nulo!")
-    private Integer idVehicle;
+    private Long idVehicle;
 
     @NotNull(message = "O cliente não pode ser nulo!")
-    private Integer idCustomer;
+    private Long idCustomer;
 }
