+++ src/main/java/com/sys/park/app/repositories/PersonRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sys.park.app.models.PersonModel;
 
-public interface PersonRepository extends JpaRepository<PersonModel, Integer> {
+public interface PersonRepository extends JpaRepository<PersonModel, Long> {
     Optional<PersonModel> findByCpf(String cpf);
     Optional<PersonModel> findByEmail(String email);
     Optional<PersonModel> findByPhone(String phone);
