+++ src/main/java/com/sys/park/app/repositories/VacancyRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.sys.park.app.models.VacancyModel;
 
 @Repository
-public interface VacancyRepository extends JpaRepository<VacancyModel, Integer> {
+public interface VacancyRepository extends JpaRepository<VacancyModel, Long> {
     Integer countBySituationTrue();
     Integer countBySituationFalse();
     List<VacancyModel> findBySituation(Boolean situation);
